Moved DHT interop example to subdir
