+++ .gitignore
@@ -1,3 +1,3 @@
 dht-interop
-js-dht-test/node_modules/
+content-dht-provide-find/js-dht-test/node_modules/
 util/private-key-gen
