+++ content-dht-provide-find/util/Makefile
@@ -1,0 +1,5 @@
+SRC		= private-key-gen.go
+BIN		= $(SRC:.go=)
+
+all: $(SRC)
+	go build -o $(BIN) $(SRC)
