--- util/Makefile
@@ -1,5 +1,0 @@
-SRC		= private-key-gen.go
-BIN		= $(SRC:.go=)
-
-all: $(SRC)
-	go build -o $(BIN) $(SRC)
